Explicitly send image data and path to PDF build process.
This is intended to make the PDF build process completely independent
of the global state database; all content required to build a test is
sent explicitly to the build process by the ProcessPoolExecutor.
Increased isolation of the build processes may simplify later API changes
addressing what happens when the main script is reimported upon process
pool creation.

The addition of a process initializer function was to address Pylint
warnings about too many arguments and instance variables, so the
initializer was created to handle common content instead of being passed
and stored as instance data.